Add PhoneFilter
Move PhoneFilter from CensusDataFilter to specific filter.
